requirements.txt
pyTelegramBotAPI==4.26.0
google-generativeai==0.9.0
gTTS
SpeechRecognition
PyAudio
python-dotenv
